Enable GPU field to be set in CLI when creating new container apps
