+++ src/containerapp/azext_containerapp/tests/latest/test_containerapp_workload_profile_commands.py
@@ -437,42 +437,3 @@ class ContainerAppWorkloadProfilesTest(ScenarioTest):
             JMESPathCheck("properties.template.containers[0].resources.cpu", cpu),
             JMESPathCheck("properties.template.containers[0].resources.memory", mem)
         ])
-
-
-class ContainerAppWorkloadProfilesGPUTest(ScenarioTest):
-    def __init__(self, *arg, **kwargs):
-        super().__init__(*arg, random_config_dir=True, **kwargs)
-
-    @AllowLargeResponse(8192)
-    @ResourceGroupPreparer(location="northeurope")
-    def test_containerapp_create_enable_dedicated_gpu(self, resource_group):
-        self.cmd('configure --defaults location={}'.format("northeurope"))
-        env = self.create_random_name(prefix='gpu-env', length=24)
-        gpu_default_name = "gpu"
-        gpu_default_type = "NC24-A100"
-        self.cmd('containerapp env create -g {} -n {} --logs-destination none --enable-dedicated-gpu'.format(
-            resource_group, env), expect_failure=False, checks=[
-            JMESPathCheck("name", env),
-            JMESPathCheck("properties.provisioningState", "Succeeded"),
-            JMESPathCheck("length(properties.workloadProfiles)", 2),
-            JMESPathCheck('properties.workloadProfiles[0].name', "Consumption", case_sensitive=False),
-            JMESPathCheck('properties.workloadProfiles[0].workloadProfileType', "Consumption", case_sensitive=False),
-            JMESPathCheck('properties.workloadProfiles[1].name', gpu_default_name, case_sensitive=False),
-            JMESPathCheck('properties.workloadProfiles[1].workloadProfileType', gpu_default_type, case_sensitive=False),
-            JMESPathCheck('properties.workloadProfiles[1].maximumCount', 1),
-            JMESPathCheck('properties.workloadProfiles[1].minimumCount', 0),
-        ])
-        containerapp_env = self.cmd('containerapp env show -g {} -n {}'.format(resource_group, env)).get_output_in_json()
-
-        while containerapp_env["properties"]["provisioningState"].lower() == "waiting":
-            time.sleep(5)
-            containerapp_env = self.cmd('containerapp env show -g {} -n {}'.format(resource_group, env)).get_output_in_json()
-        app1 = self.create_random_name(prefix='app1', length=24)
-        self.cmd(f'containerapp create -n {app1} -g {resource_group} --image mcr.microsoft.com/azuredocs/samples-tf-mnist-demo:gpu --environment {env} -w {gpu_default_name} --min-replicas 1 --cpu 0.1 --memory 0.1', checks=[
-            JMESPathCheck("properties.provisioningState", "Succeeded"),
-            JMESPathCheck("properties.workloadProfileName", gpu_default_name),
-            JMESPathCheck('properties.template.containers[0].resources.cpu', '0.1'),
-            JMESPathCheck('properties.template.containers[0].resources.memory', '0.1Gi'),
-            JMESPathCheck('properties.template.scale.minReplicas', '1'),
-            JMESPathCheck('properties.template.scale.maxReplicas', '10')
-        ])
